modified calc

+++ assignments/JS/SPA/controller/CustomerController.js
@@ -6,23 +6,27 @@ let cusName = $('#txtCusName');
 let cusAddress = $('#txtCusAddress');
 let cusSalary = $('#txtCusSalary');
 
+$('#btnSave').attr("disabled", true);
+
 $('#btnSave').click(function (event){
+    var userChoice = window.confirm("Do you want to save the customer?");
 
-    customer.addValue(cusId.val(), cusName.val(), cusAddress.val(), cusSalary.val());
+    if (userChoice) {
+        customer.addValue(cusId.val(), cusName.val(), cusAddress.val(), cusSalary.val());
+        customerDetail.push(customer);
 
-    customerDetail.push(customer);
-
-    $('#cusTBody').append(
-        `<tr>
+        $('#cusTBody').append(
+            `<tr>
             <th scope="row">${customer.id}</th>
             <td>${customer.name}</td>
             <td>${customer.address}</td>
             <td>${customer.salary}</td>
             <td style="width: 10%"><img class="delete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
         </tr>`
-    );
-    deleteDetail();
-    setFeilds();
+        );
+        deleteDetail();
+        setFeilds();
+    }
     event.preventDefault();
 })
 
@@ -35,10 +39,15 @@ $('#clear').click(function (event){
 })
 
 $('#getAll').click(function (){
+    getAll();
+
+})
+
+function getAll() {
     let tBody = $('#cusTBody')
     tBody.empty();
 
-    for (let i = 0; i <customerDetail.length ; i++) {
+    for (let i = 0; i < customerDetail.length; i++) {
         tBody.append(`<tr>
             <th scope="row">${customerDetail[i].id}</th>
             <td>${customerDetail[i].name}</td>
@@ -48,9 +57,9 @@ $('#getAll').click(function (){
             </tr>`);
         deleteDetail();
         setFeilds();
-    };
-
-})
+    }
+    ;
+}
 
 setFeilds();
 
@@ -60,6 +69,7 @@ function setFeilds() {
         cusName.val($(this).children(':eq(1)').text());
         cusAddress.val($(this).children(':eq(2)').text());
         cusSalary.val($(this).children(':eq(3)').text());
+        $('#btnSave').attr("disabled", true);
     })
 }
 
@@ -72,19 +82,26 @@ function deleteDetail() {
     )
 
     btnDelete.click(function () {
-        $(this).parents('tr').remove();
+        var userChoice = window.confirm("Do you want to delete the customer?");
+
+        if (userChoice) {
+            $(this).parents('tr').remove();
+        }
     })
 }
 
 $('#btnSearch').click(function (){
     let id = $('#txtSearch').val();
     let tbody = $('#cusTBody');
+    let count = 0;
 
-    for (let i = 0; i < customerDetail.length; i++) {
-        if(customerDetail[i].id==id){
-            tbody.empty();
+    if(id.length!=0) {
+        for (let i = 0; i < customerDetail.length; i++) {
+            if (customerDetail[i].id == id) {
+                count++;
+                tbody.empty();
 
-            tbody.append(`<tr>
+                tbody.append(`<tr>
                 <th scope="row">${customerDetail[i].id}</th>
                 <td>${customerDetail[i].name}</td>
                 <td>${customerDetail[i].address}</td>
@@ -93,6 +110,24 @@ $('#btnSearch').click(function (){
                 </tr>`);
                 deleteDetail();
                 setFeilds();
+            }
+        }
+        if (count != 1) {
+            alert("No such Customer..please check the ID");
+        }
+    }else {
+        alert("Please enter the ID");
+    }
+})
+
+$('#edit').click(function (){
+    for (let i = 0; i < customerDetail.length; i++) {
+
+        if(customerDetail[i].id == $('#txtCusID').val()){
+            customerDetail[i].name = $('#txtCusName').val();
+            customerDetail[i].address = $('#txtCusAddress').val();
+            customerDetail[i].salary = $('#txtCusSalary').val();
+            getAll();
         }
     }
 })
