+++ assignments/JS/SPA/controller/ItemController.js
@@ -1,28 +1,32 @@
 import {item} from "../model/Item.js";
 import {itemDetail} from "../db/DB.js";
 
-let itemCode = $('#itemCode');
-let itemName = $('#itemName');
-let itemPrice = $('#itemPrice');
-let itemQuantity = $('#itemQuantity');
+let itemCode = $('#txtItemCode');
+let itemName = $('#txtItemName');
+let itemPrice = $('#txtItemPrice');
+let itemQuantity = $('#txtItemQuantity');
 
 $('#itemSave').click(function (event){
+    var userChoice = window.confirm("Do you want to save the item?");
 
-    item.addValue(itemCode.val(), itemName.val(), itemPrice.val(), itemQuantity.val());
+    if (userChoice) {
 
-    itemDetail.push(item);
+        item.addValue(itemCode.val(), itemName.val(), itemPrice.val(), itemQuantity.val());
 
-    $('#itemBody').append(
-        `<tr>
+        itemDetail.push(item);
+
+        $('#itemBody').append(
+            `<tr>
             <th scope="row">${item.code}</th>
             <td>${item.name}</td>
             <td>${item.price}</td>
             <td>${item.quantity}</td>
-            <td style="width: 10%"><img class="delete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
+            <td style="width: 10%"><img class="itemDelete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
         </tr>`
-    );
-    deleteDetail();
-    setFeilds();
+        );
+        deleteDetail();
+        setFeilds();
+    }
     event.preventDefault();
 })
 
@@ -44,7 +48,7 @@ $('#itemGetAll').click(function (){
             <td>${itemDetail[i].name}</td>
             <td>${itemDetail[i].price}</td>
             <td>${itemDetail[i].quantity}</td>
-            <td style="width: 10%"><img class="delete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
+            <td style="width: 10%"><img class="itemDelete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
             </tr>`);
         deleteDetail();
         setFeilds();
@@ -66,14 +70,19 @@ function setFeilds() {
 deleteDetail();
 
 function deleteDetail() {
-    let btnDelete = $('.delete');
+    let btnDelete = $('.itemDelete');
     btnDelete.on("mouseover", function (){
         $(this).css("cursor", "pointer");}
     )
 
     btnDelete.click(function () {
-        $(this).parents('tr').remove();
+        var userChoice = window.confirm("Do you want to delete the item?");
+
+        if (userChoice) {
+            $(this).parents('tr').remove();
+        }
     })
+
 }
 
 $('#itemSearch').click(function (){
@@ -89,7 +98,7 @@ $('#itemSearch').click(function (){
                 <td>${itemDetail[i].name}</td>
                 <td>${itemDetail[i].price}</td>
                 <td>${itemDetail[i].quantity}</td>
-                <td style="width: 10%"><img class="delete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
+                <td style="width: 10%"><img class="itemDelete" src="../../CSS_Framework/POS/assets/icons8-delete-96.png" alt="Logo" width="50%" className="opacity-75"></td>
                 </tr>`);
             deleteDetail();
             setFeilds();
